Assign selector column in CopyTable::load
The CopyTable contains a BinaryNumberChip that uses a selector, but the load method doesn't enable the selector.  The result of this is that when testing standalone circuits like the EVM circuit, the CopyTable will be included with a fixed column that is all zeroes.  This causes problems when testing the EVM circuit with the Root aggregation circuit, because it doesn't support zeroed fixed columns.  This commit should fix the problem.